+++ javascript/node/selenium-webdriver/lib/webdriver.js
@@ -2254,6 +2254,23 @@ class Alert {
   }
 
   /**
+   * Sets the username and password in an alert prompting for credentials (such
+   * as a Basic HTTP Auth prompt). This method will implicitly
+   * {@linkplain #accept() submit} the dialog.
+   *
+   * @param {string} username The username to send.
+   * @param {string} password The password to send.
+   * @return {!promise.Promise<void>} A promise that will be resolved when this
+   *     command has completed.
+   */
+  authenticateAs(username, password) {
+    return this.driver_.schedule(
+        new command.Command(command.Name.SET_ALERT_CREDENTIALS),
+        'WebDriver.switchTo().alert()'
+            + `.authenticateAs("${username}", "${password}")`);
+  }
+
+  /**
    * Accepts this alert.
    *
    * @return {!promise.Promise<void>} A promise that will be resolved
@@ -2344,6 +2361,16 @@ class AlertPromise extends Alert {
     this.getText = function() {
       return alert.then(function(alert) {
         return alert.getText();
+      });
+    };
+
+    /**
+     * Defers action until the alert has been located.
+     * @override
+     */
+    this.authenticateAs = function(username, password) {
+      return alert.then(function(alert) {
+        return alert.authenticateAs(username, password);
       });
     };